- Assume also needed real8 for 'zsq' in subroutine read_nc_ini_file(), as it is directly compared to 'inizs', which would now be real8

+++ source/src/sfincs_initial_conditions.F90
@@ -261,7 +261,7 @@ contains
       !
       implicit none
       !
-      real*4, dimension(:),   allocatable :: zsq
+      real*8, dimension(:),   allocatable :: zsq
       integer   :: npq, ip
       !
       NF90(nf90_open(trim(ncinifile), NF90_CLOBBER, net_file_ini%ncid))
